at-spi2-core: use dbus-broker if available
The at-spi2-core now has upstream support for alternative
bus-implementations, mainly 'dbus-broker' over 'dbus-daemon'. If a path
to either binary is provided at compile time, at-spi2 will try the
given bus implementation at runtime. Since Fedora provides two
alternatives, we make at-spi2 use both. Note that at-spi2 falls back to
the alternative if the primary pick is not available.

We keep the 'Requires: dbus' dependency. The 'dbus' package was turned
into a meta-package. Right now it pulls in either 'dbus-daemon' or
'dbus-broker', but in the future it will be a real meta-package and be
'Provides: dbus' by both dbus-daemon and dbus-broker. Hence, the
depdency is sufficient to guarantee that at least one bus
implementation is available on the machine.

Given that dbus-broker will be the default bus implementation in F29,
this patch also switches the at-spi2 default to 'dbus-broker'. Again,
this will safely fall-back to 'dbus-daemon' at runtime if dbus-broker
is unavailable.

(dvdhrm: This is based on zbyszek's original patch, but now highly
         modified by me.)

v2:
 - fix __linux__ vs __linux preprocessor define

v3:
 - backport the __linux__ fix to 2.28.0 to make it apply cleanly